+++ src/main/main.css
@@ -1,10 +1,10 @@
-.MainP {
+.main {
   font-family: Arial, sans-serif;
   padding: 20px;
 }
 
 .logo-header {
-  font-size: 24px; 
+  font-size: 24px;
 }
 
 .header {
@@ -16,9 +16,7 @@
 
 .header img {
   width: 110px;
-
   height: auto;
-
 }
 
 h1 {
@@ -29,9 +27,7 @@ h1 {
   display: flex;
   align-items: center;
   justify-content: center;
-
   flex: 1;
-
 }
 
 input[type='text'] {
